Vulkan compute plugin, SPIR-V emitter, u64 SignalBits, CLI cleanup
GPU backend (plugins/vulkan/):
- Vulkan compute dispatch with ash + gpu-allocator
- Three-phase async: dispatch (non-blocking submit), wait (fiber
  suspends on fence fd via VK_KHR_external_fence_fd + IoOp::PollFd),
  collect (readback). Zero thread pool threads consumed.
- vulkan/submit convenience wraps all three for simple cases

SPIR-V emitter (lib/spirv.lisp):
- Runtime SPIR-V bytecode generation from Elle code
- No GLSL, no glslc, no shaderc — Elle generates GPU bytecode directly
- Supports f32 storage buffers, arithmetic, comparisons, select, u2f/f2u

GPU library (lib/gpu.lisp):
- gpu:compile wires SPIR-V emitter to Vulkan pipeline creation
- gpu:run does dispatch/wait/collect with decoded f32 results

SignalBits u32 → u64:
- Full 64-bit width in bytecode encoding (four u16s), JIT, WASM ABI
- WASM signal locals, returns, host functions all widened to i64
- No truncation of user-defined signals

CLI cleanup:
- --help shows CLI args only, no primitives list
- ELLE_HOME/ELLE_PATH/ELLE_CACHE as env var defaults, overridden by
  --home=/--path=/--cache= CLI args
- Module resolution uses config (CLI > env var > default)

Core visibility:
- IoRequest, IoOp, and related types made pub for plugin async I/O

+++ src/wasm/host.rs
@@ -54,7 +54,7 @@ pub struct WasmSuspensionFrame {
     pub env_base: usize,
     /// Full signal bits at the yield point. Preserves SIG_IO and other
     /// bits so the scheduler can detect I/O requests on the fiber.
-    pub signal_bits: u32,
+    pub signal_bits: u64,
 }
 
 /// Host state stored in the Wasmtime `Store<ElleHost>`.
